App/BackEnd/sentiment-analysis-manager/requirements.txt
python-dotenv==1.0.1
requests==2.31.0
pandas==2.0.3
#numpy==1.24.4
colorlog==6.8.2